backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
supabase==2.3.4
pdfplumber==0.11.0
google-generativeai==0.8.3
python-dotenv==1.0.1

# Google OAuth + Gmail API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# Email notifications
resend==0.7.2
